# docs/swagger.yaml
basePath: /
definitions:
  handler.errorMessage:
    properties:
      message:
        type: string
    type: object
  handler.getAllListsResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/todo_app.TodoList'
        type: array
    type: object
  handler.signInInput:
    properties:
      password:
        type: string
      username:
        type: string
    required:
    - password
    - username
    type: object
  handler.statusResponse:
    properties:
      status:
        type: string
    type: object
  todo_app.ListsItem:
    properties:
      id:
        type: integer
      itemId:
        type: integer
      listId:
        type: integer
    type: object
  todo_app.TodoItem:
    properties:
      description:
        type: string
      done:
        type: boolean
      id:
        type: integer
      title:
        type: string
    required:
    - title
    type: object
  todo_app.TodoList:
    properties:
      description:
        type: string
      id:
        type: integer
      title:
        type: string
    required:
    - title
    type: object
  todo_app.UpdateItemInput:
    properties:
      description:
        type: string
      done:
        type: boolean
      title:
        type: string
    type: object
  todo_app.UpdateListInput:
    properties:
      description:
        type: string
      title:
        type: string
    type: object
  todo_app.User:
    properties:
      name:
        type: string
      password:
        type: string
      username:
        type: string
    required:
    - name
    - password
    - username
    type: object
host: localhost:8000
info:
  contact: {}
  description: Api Server for TodoList Application
  title: Todo App Api
  version: "1.0"
paths:
  /api/items/:id:
    delete:
      consumes:
      - application/json
      description: delete item
      operationId: delete-item
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.statusResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.errorMessage'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.errorMessage'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.errorMessage'
        default:
          description: ""
          schema:
            $ref: '#/definitions/handler.errorMessage'
      security:
      - ApiKeyAuth: []
      summary: Delete Item
      tags:
      - items
    get:
      consumes:
      - application/json
      description: get item by id
      operationId: get-item-by-id
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/todo_app.TodoItem'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.errorMessage'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.errorMessage'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.errorMessage'
        default:
          description: ""
          schema:
            $ref: '#/definitions/handler.errorMessage'
      security:
      - ApiKeyAuth: []
      summary: Get Item By Id
      tags:
      - items
    put:
      consumes:
      - application/json
      description: update item
      operationId: update-item
      parameters:
      - description: update item
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/todo_app.UpdateItemInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.statusResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.errorMessage'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.errorMessage'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.errorMessage'
        default:
          description: ""
          schema:
            $ref: '#/definitions/handler.errorMessage'
      security:
      - ApiKeyAuth: []
      summary: Update Item
      tags:
      - items
  /api/lists:
    get:
      consumes:
      - application/json
      description: get all lists
      operationId: get-all-lists
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.getAllListsResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.errorMessage'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.errorMessage'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.errorMessage'
        default:
          description: ""
          schema:
            $ref: '#/definitions/handler.errorMessage'
      security:
      - ApiKeyAuth: []
      summary: Get All Lists
      tags:
      - lists
    post:
      consumes:
      - application/json
      description: create todo_app list
      operationId: create-list
      parameters:
      - description: list info
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/todo_app.TodoList'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: integer
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.errorMessage'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.errorMessage'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.errorMessage'
        default:
          description: ""
          schema:
            $ref: '#/definitions/handler.errorMessage'
      security:
      - ApiKeyAuth: []
      summary: Create todo_app list
      tags:
      - lists
  /api/lists/:id:
    delete:
      consumes:
      - application/json
      description: delete list
      operationId: delete-list
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.statusResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.errorMessage'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.errorMessage'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.errorMessage'
        default:
          description: ""
          schema:
            $ref: '#/definitions/handler.errorMessage'
      security:
      - ApiKeyAuth: []
      summary: Delete List
      tags:
      - lists
    get:
      consumes:
      - application/json
      description: get list by id
      operationId: get-list-by-id
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/todo_app.ListsItem'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.errorMessage'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.errorMessage'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.errorMessage'
        default:
          description: ""
          schema:
            $ref: '#/definitions/handler.errorMessage'
      security:
      - ApiKeyAuth: []
      summary: Get List By Id
      tags:
      - lists
    put:
      consumes:
      - application/json
      description: update list
      operationId: update-list
      parameters:
      - description: update list
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/todo_app.UpdateListInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.statusResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.errorMessage'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.errorMessage'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.errorMessage'
        default:
          description: ""
          schema:
            $ref: '#/definitions/handler.errorMessage'
      security:
      - ApiKeyAuth: []
      summary: Update List
      tags:
      - lists
  /api/lists/:id/items:
    get:
      consumes:
      - application/json
      description: get todo_app items
      operationId: get-all-items
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: integer
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.errorMessage'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.errorMessage'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.errorMessage'
        default:
          description: ""
          schema:
            $ref: '#/definitions/handler.errorMessage'
      security:
      - ApiKeyAuth: []
      summary: Get todo_app items
      tags:
      - items
    post:
      consumes:
      - application/json
      description: create todo_app item
      operationId: create-item
      parameters:
      - description: item info
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/todo_app.TodoItem'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: integer
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.errorMessage'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.errorMessage'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.errorMessage'
        default:
          description: ""
          schema:
            $ref: '#/definitions/handler.errorMessage'
      security:
      - ApiKeyAuth: []
      summary: Create todo_app item
      tags:
      - items
  /auth/sign-in:
    post:
      consumes:
      - application/json
      description: login
      operationId: login
      parameters:
      - description: credentials
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/handler.signInInput'
      produces:
      - application/json
      responses:
        "200":
          description: token
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.errorMessage'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.errorMessage'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.errorMessage'
        default:
          description: ""
          schema:
            $ref: '#/definitions/handler.errorMessage'
      summary: SignIn
      tags:
      - auth
  /auth/sign-up:
    post:
      consumes:
      - application/json
      description: create account
      operationId: create-account
      parameters:
      - description: account info
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/todo_app.User'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: integer
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.errorMessage'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.errorMessage'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.errorMessage'
        default:
          description: ""
          schema:
            $ref: '#/definitions/handler.errorMessage'
      summary: SignUp
      tags:
      - auth
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"